implement destroy for a single locale

+++ app/models/data_cycle_core/content/destroy_content.rb
@@ -3,7 +3,8 @@
 module DataCycleCore
   module Content
     module DestroyContent
-      def destroy_content(current_user: nil, save_time: Time.zone.now, save_history: true, destroy_linked: false)
+      def destroy_content(current_user: nil, save_time: Time.zone.now, save_history: true, destroy_locale: false, destroy_linked: false)
+        return if destroy_locale && !available_locales.include?(I18n.locale)
         ActiveRecord::Base.transaction do
           children.each { |item| item.destroy_content(current_user: current_user, save_time: save_time) } if respond_to?(:children)
           unless history? || !save_history
@@ -11,17 +12,27 @@ module DataCycleCore
             self.deleted_by = current_user&.id
             to_history(save_time: save_time, delete: true)
           end
-          destroy_children(current_user: current_user, save_time: save_time, destroy_linked: destroy_linked)
+
+          destroy_children(current_user: current_user, save_time: save_time, destroy_linked: destroy_linked, destroy_locale: destroy_locale)
           destroy_linked_data(current_user: current_user, save_time: save_time, save_history: save_history, destroy_linked: destroy_linked) if destroy_linked
-          destroy
+          if destroy_locale
+            destroy_item(self, (available_locales - [I18n.locale]).blank?)
+          else
+            destroy
+          end
         end
         run_callbacks(:destroyed_data_hash) unless history?
       end
 
-      def destroy_children(current_user: nil, save_time: Time.zone.now, destroy_linked: false)
+      def destroy_children(current_user: nil, save_time: Time.zone.now, destroy_linked: false, destroy_locale: false)
         embedded_property_names.each do |name|
-          load_embedded_objects(name, false).each do |item|
-            item.destroy_content(current_user: current_user, save_time: save_time, save_history: false, destroy_linked: destroy_linked)
+          load_embedded_objects(name, destroy_locale).each do |item|
+            if destroy_locale
+              item.destroy_children(current_user: current_user, save_time: save_time, destroy_linked: destroy_linked, destroy_locale: destroy_locale)
+              destroy_item(item, (item.available_locales - [I18n.locale]).blank?)
+            else
+              item.destroy_content(current_user: current_user, save_time: save_time, save_history: false, destroy_linked: destroy_linked, destroy_locale: destroy_locale)
+            end
           end
         end
         asset_property_names.each do |name|
@@ -43,6 +54,15 @@ module DataCycleCore
           DataCycleCore::ContentContent.where(content_a_id: item_id).pluck(:content_b_id) +
           DataCycleCore::ContentContent.where(content_b_id: item_id).pluck(:content_a_id)
         ).uniq.size
+      end
+
+      def destroy_item(item, condition)
+        if condition
+          item.destroy
+        else
+          item.translation.destroy
+          item.translations.reload
+        end
       end
     end
   end
